hack for mail-template
if a link is empty, the opening link tag is not rendered but the closing tag
is still here because of the insufficient regex of the getContentParts() method
in the text row. if this occurs, the mail html parser quits. Correct solution
would be to check this in getContentParts(), this is the hack version.