Make model settable through settings
This is needed because to implement a test you will need to use a fnf model which is now possible.
In Favourites/Page/Component the model is set explicitly because it's not dependent of favourites/component, because favourites/component is just the component added to another component to favourise it.