Enable setting cssClass for fieldset.
This is needed when you have more fieldset in a single form which behave differently.
The concrete case was having 3 normal fieldset and one with default collapse-behaviour
but problem to fit in layout.
Using setCls because this is how TextField does and how ext calls the property for
Ext.form.FieldSet.