Fix problem with copy of page from one domain into another
The problem was that there where different root_category components with a different class
and the old code cleared only the cache of the source root_category component but not of
the target component. Now the cache of the target root_category component is cleared and
the new copied componend is found.