Change bestfit-paramname to cover
This parameter-name is more clearly. If cover parameter is set to true
the provided width and height have to be covered. So if the image is
to small and has a different aspect ratio it gets scaled up and cut off
to match the aspect ratio. (equal to bestfit=false before this change)
If cover parameter is set to false the image can be smaller or stay
in different aspect ratio than provided width and height.