Fix problem with responsive lightbox and adaptHeight option
If lightbox was opened after clicking link height was not set correctly,
because it was calculated differently after content was loaded. It
actually worked when reloading the lightbox. Now using the same code
for both situations.
Was not possible to directly call the other code because some elements
weren't set/existing.