Use CURDATE instead of NOW if sql comparing with DATE-column
MySQL changed logic how to compare a DATE-column with NOW() from 5.0 to 5.1
It's recommended to compare DATE-column with CURDATE() instead of NOW() to
be uneffected by this change. This is a problem with our production-online
version of MySQL and our local work-environment.
See http://bugs.mysql.com/bug.php?id=28929 for more information.