Work around Php 5.2 crash
Woohoo,
after days of bad sleeping and after several hours of debugging we finally
could locate 'is_subclass_of' as problematic function call.

https://bugs.php.net/bug.php?id=50360

I hope this is the actual cause of the random crashes we get.

Bug has been fixed in Php 5.3 - but we are stuck with 5.2 - Danke POI... für nix