break loop when no parentNode exists
this can happen if dom element was removed from document and is still in some caches

owl-carousel does this:
in some cases (probably when owl option changes dynamically and refreshes) owl deletes
some items while a "while loop" is running in "doesElementScroll()" who uses one of that items.
Then the element in the while-loop becomes null while "doesElementScroll" trys to get the parentNode
from that deleted element and then we get an error.