Don't use static array to define fieldKey. Give option to set manually
Problem on page with multiple forms containing multicheckbox for same
relationToValuesRule. Results in deleting relation-rows in submited
form (no problem with the first form on page) after changing values.
The reason was that on page load every form is initialised and therefore
fieldKey for multicheckbox#2 is form_RELATION2_ID. In post-request only
posted form ist initialised and therefore multicheckbox in form#2 is called
form_RELATION_ID and does not match with post-variables with value
form_RELATION2_ID