Union Model: Implement selecting from sibling model
Implemented not very efficient, as with one query all ids are fetched from sibling and passed to actual union query.

TODO: Execute just a single query if all models are Db models.
(That is not that easily possible as we would have to adjust Model_Db which should be better done in master branch)