Remove code generating setlocale-part for setup.php
This can be simply removed because it doesn't work in multilingual
webs or translated backend. Even if it is a untranslated web we
should be working with component-data for locale-formating. For
any old webs depending on setlocale it can be easily set in bootstrap.
This also resulted in problems with trl-build because generating
the setlocale-part of setup.php required translation which isn't
possible before build which required generated setup.php.