Fix column component responsive bug
When a column component with properties 25-25-25-25 is used and
its parent is between 480 and 621 pixels wide, than it looks like a two column component.
Problem is: when the content of the first element is higher than the second one,
than the third element will move to the right edge, because of an missing "clear" setting in css.
This commit should fix this wrong behavior.