Add test for fullpage clearing
This test asserts implementation details (like how many entries have been deleted) - which should not be done - but in this case we want to test if not too many entries get deleted.

When deleting a component cache entry the includes of type componentLink don't need to be followed. (bad for performance)