Change highlighting of searchresults and searchresults text itself
Change highlightingTerms function to always get resulttext with the searchterms in it.
Change highlighting to hihglight all partials of the searchterms, not only the exact alone standing word.
This behavior can be changed with the "highlightWordPartial" option.

Add a Test to check if the resulttext has the searchterms in it.