Use same identifier for onElementHide as for onElementShow in Youtube Component
When having more Youtube Lightbox Links on one page closing the second window
did not cause the video to stop. Problem was that on the first closing the first
lightbox was saved in elCacheBySelector and on the second closing the cache was
not empty and assumed to contain the second lightbox which wasn't true.