Improved trl duplicate exceptions during build.
Problem 1: translation is defined in several packages but translated
differently. In this case there is no way to define which translation
should be used. So somebody needs to fix this conflict by either
translate it exactly the same or add context to one or both.
Problem 2: translation is defined in kwf and other packages. Only
kwf translation is used. This shouldn't be possible if packages
where parsed via kwf-trl-parser and therefore this should never happen.