Defer Admin javascript
Fixes problems when dependency is used in Deferred Frontend Package and Admin. It had to be loaded twice which caused problems for some assets.

This defines a Kwf.main entry point for the application that will be called at the end of Admin js.

This also adds a AdminMain dependency that will be used by Admin package.