Change AccessDenied to 403, add 401 Unauthorized Exception
Throw correct exception with additional message in setup.php.
401 is used if user can't be authorized (e.g. missing, wrong
or expired credentials). This can happen anytime and needs
to be handled correctly or displayed to the user.
403 is used if user is authorized but not allowed to access
or do whatever he wanted to do. This should never happen in
normal application flow.
This is an incompatible change but can be neglected because
we don't handle exceptions considering their status code in
web and didn't handle in native applications.
From now on our native applications do check status code to
handle unauthorized with refreshing token or promt the user
to login.