kwf/check: Remove check for db connection
This url is used to check if the webserver is up and running (and can be used by load balancer).
When the database is down that's a different issue the webserver is not responsible for, and
can not be solved by restarting the webserver or something.

Improves OpenShift restarting behaviour