Fix possible issue with view cache entries getting seized up in memcache
This happened when a fatal error happend during rendering (eg. a php runtime timeout) where
the writeBuffer methods where not called (as the __destruct of the model was
not called.

It was not a problem when calling exit manually, as that triggers all destructors. (How awesome, Php)

With this change the buffer is now disabled and view cache entries are saved into mysql
even before setting them in memcache.

The buffer existed for performance reasons (to bulk insert multiple entries), so with this
change saving gets slightly slower. But under normal circumstantes we don't save that often.