Update search results if query is empty
When typing into the search field the form would not automatically
start searching, if the length of the search query is less than three
characters long.
Therefore when all the text is delete from the search field the search
results would not be updated and would show the results of the last
query until the user hits 'enter'.

Now the search results will also be updated,
when the search query is empty.