Directory_List change: move getItems/getItemIds from View to List
It should always have been like that.

View is responsible for output/styling, List is responsible for getting the data.
With that change more related to getting the data is moved where it should be.

This allows overriding getItems and implementing own Detail pages under a List, with
that the detail can have custom content per List.

This change is obviously very incompatible but hopefully not too many Views did override getItems/getItemIds.