Fix for wrong topoffset on success
If this.config.hideFormOnSuccess is set to true, this.el will be set to display: none; and therefore it has no topoffset anymore. For example, if the form is in the middle of the page and sending via ajax is complete, it will scroll to the top off the page, because topoffset is 0. But what we realy want is to scroll to the top of the form, to see the successmessage.