Optimize the dropdown menu for touch-devices
Previously when hovering the cursor over a link with a dropdown menu,
the dropdown would not appear on devices that support touch events.
When tapping on the link the dropdown menu would not appear
because the link would open immediately.

Using DoubleTapToGo now allows the user to open the dropdown menu
before opening the link when tapping and the dropdown menu is also
shown when hovering the cursor over the link.