Forward redirect parameter from registration-form to activation-form
This ways it's possible to allow to continue where the user started.
RedirectUrl forward is implemented like in login-form.
RedirectUrl has to be set in User_EditRow because sendActivationMail is
called in save() function of User_EditRow and not directly in registration
component. It also needs to be an additional parameter for
Kwf_User_Mail_Activation because $row is no User_EditRow and does not have
the redirectUrl field.