Implement possibility to add maintenance jobs by defining a tagged symfony service
Tag service as kwf.maintenance_job and it will be picked up as maintenance job.

Also add a new Base Class for Symfony MaintenanceJobs that uses PSR-3 for logging and use monolog
when running the job. For this the execute signature has to be changed which is not possible, to hack
around that I moved Kwf/Util/Maintenance/Job/Abstract.php to Kwf/Util/Maintenance/Job/AbstractBase.php
which has no execute method and added the execute() in Kwf/Util/Maintenance/Job/Abstract.php
and execute(LoggerInterface $logger) in KwfBundle/MaintenanceJobs/AbstractJob.php