Delete files and folders when cleaning the build directory
The koala build folder can deal with files only. When integrating decoupled child projects into a project built with koala-framework there is a need to move folders with files (subfolders) into the build folder. When the build folder is cleaned during the build process, subfolders and files in the build folder get now successfully removed