Implement LoginRedirect not in View Plugin
Using ViewReplace plugin for a redirect (using header('Location... and exit) never has been a good idea.
It is a practical problem now with ContentApi where we get the redirect for the api request.

This change adds a new Redirect Plugin Interface that allows a plugin returning a redirect url. This is done before
rendering and processInput (which will solve a few problems we had there with unauthenticated users)

Downside of this change: it only works when the plugin is at page level