Add redirect-url to params
This was done to let the login-page know the original url when a user tries
to login via facebook and get's redirected to the login-page again.
A popup will be shown there that requires the user to accept terms and conditions
and only on acceptance will a user be redirected to the actual page they came from.