Fix duplicating statically generated Text Component when target already exists in database
Fixes issue with component-deep-copy when Text is generated statically and for some reason the row already exists. Other components (Kwc_Abstract_Admin::_duplicateOwnRow) do the same.

Also remove completely broken _beforeDelete for Text Row. The implementation tried to delete child components using Kwc_Admin::delete which never really existed. But the
first problem was already in instanciating the ChildComponentsModel (as that requires a constructor parameter) - so this was never called and can be safely removed.