Update Kwf_Util_Geocode to use api-key for google-geocode-requests
A new config setting googleServerToServerApiKey is introduced to handle
server to server requests (e.g Google Geocode API or other APIs that don't
support http-referrer restrictions). This key shall not not be exposed to
the client because it might not be restricted. For client-exposed keys the
config setting googleMapsApiKey remains.