Remove window.open in download-tag
This behaviour was to prevent opening the download directly in the
current browser window (even though the we tell the browser to
download the file anyway). Nowadays this workaround should not be
necessary anymore.
Current problem is in mobile-apps when the link is clicked in an
in-app-browser because an empty webview is opened.