Add fallback for redirect-param
An infinite loop occurs if the redirect-param is not set. The user is being redirected
to the login-page where the authorization request is sent multiple times with the same
access-token, which does not work because the access-token has already been invalidated.
Redirecting to the home-page works because at this point the user is already loggend in.